Organize archiver into module
