+++ src/nimbuscli/cmd/factory.py
@@ -9,13 +9,8 @@ from nimbuscli.cmd.backup import Backup
 from nimbuscli.cmd.command import Command
 from nimbuscli.cmd.deploy import Down, Up
 from nimbuscli.config import Config
-from nimbuscli.core import (
-    Archiver,
-    AwsUploader,
-    RarArchiver,
-    SubprocessRunner,
-    Uploader,
-)
+from nimbuscli.core import AwsUploader, SubprocessRunner, Uploader
+from nimbuscli.core.archive import Archiver, RarArchiver
 from nimbuscli.provider import (
     DirectoryProvider,
     Secrets,
