+++ src/nimbuscli/core/__init__.py
@@ -1,4 +1,4 @@
-from nimbuscli.core.archiver import ArchivalStatus, Archiver, RarArchiver
+from nimbuscli.core.archive import ArchivalStatus, Archiver, RarArchiver
 from nimbuscli.core.runner import CompletedProcess, Runner, SubprocessRunner
 from nimbuscli.core.service import DockerService, OperationStatus, Service
 from nimbuscli.core.uploader import AwsUploader, Uploader, UploadProgress, UploadStatus
