+++ src/nimbuscli/core/archive/__init__.py
@@ -1,0 +1,2 @@
+from nimbuscli.core.archive.archiver import ArchivalStatus, Archiver
+from nimbuscli.core.archive.rar import RarArchiver
