+++ tests/core/archive/test_rar.py
@@ -1,7 +1,7 @@
 import pytest
-from mock import Mock, PropertyMock, call, patch
+from mock import Mock, call, patch
 
-from nimbuscli.core import ArchivalStatus, RarArchiver
+from nimbuscli.core.archive import RarArchiver
 
 
 class TestRarArchiver:
@@ -116,27 +116,3 @@ class TestRarArchiver:
             assert f"-rr{recovery}" not in cmd
         else:
             assert f"-rr{recovery}" in cmd
-
-
-class TestArchivalStatus:
-
-    @pytest.mark.parametrize("folder", [True, False])
-    @pytest.mark.parametrize("archive", [True, False])
-    @pytest.mark.parametrize("success", [True, False])
-    @pytest.mark.parametrize("exists", [True, False])
-    def test_success(self, folder, archive, success, exists):
-        patcher = patch("os.path.exists")
-        mock_exists = patcher.start()
-        mock_exists.return_value = exists
-
-        mock_proc = Mock()
-        success_mock = PropertyMock(return_value=success)
-        type(mock_proc).success = success_mock
-
-        fdr = "folder" if folder else None
-        arc = "archive" if archive else None
-
-        a = ArchivalStatus(mock_proc, fdr, arc)
-
-        assert a.success == all([folder, archive, success, exists])
-        success_mock.assert_called_once()
